AUG 25/Data processing 2/requirements.txt
pandas
azure-storage-blob
